requirements.txt
fastapi
uvicorn
python-multipart
openai
faiss-cpu
tiktoken
python-dotenv
langchain
networkx